допилить фильтрацию
